+++ libs/geshi/geshi/geshi/go.php
@@ -324,10 +324,14 @@ $language_data = array(
             0 => 'color: #cc66cc;'
             ),
         'NUMBERS' => array(
-            0 => 'color: #cc66cc;'
+            0 => 'color: #cc66cc;',
+            1 => 'color: #cc66cc;',
+            2 => 'color: #cc66cc;'
             ),
         'METHODS' => array(
-            0 => 'color: #004000;'
+            0 => 'color: #004000;',
+            1 => 'color: #004000;',
+            2 => 'color: #004000;'
             ),
         'SYMBOLS' => array(
             # delimiters