+++ CHANGELOG.markdown
@@ -21,6 +21,7 @@ Frozen changes
 ### 1.0.0 (2013-06-03)
 
 * BUG: Fixed missing new line after a Geshi node.
+* BUG: Fixed crash in Geshi with "Go" language.
 
 ### 1.0.0-rc4 (2013-05-31)